% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPredictions.R
\name{build_prediction}
\alias{build_prediction}
\title{Build Predictions with Preprocessed Data Frames}
\usage{
build_prediction(query_len, hsp_table)
}
\arguments{
\item{query_len}{Length of the Query}

\item{hsp_table}{Table of HSP values}
}
\description{
Sample Description
}
