% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPredictions.R
\name{build_predictions}
\alias{build_predictions}
\title{Prediction wrapper}
\usage{
build_predictions(raw_carrier)
}
\arguments{
\item{raw_carrier}{raw carrier preprocessed from before}
}
\description{
sample description
}
