% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePredictionStats.R
\name{calculate_coverage}
\alias{calculate_coverage}
\title{Calculate Coding Sequence Coverage}
\usage{
calculate_coverage(seq)
}
\arguments{
\item{seq}{The predicted coding sequence to get sequence coverage for.}
}
\value{
The coding sequence coverage of the predicted sequence, as a numeric
value
}
\description{
Calculate the percent of the subject species' coding sequence that was
successfully predicted for a particular gene. Does so by determining the
number of gaps in the sequence (corresponding to missing sections), then
dividing the remaining nucleotides by the total length.
}
\references{
Wickham H (2023). stringr: Simple, Consistent Wrappers for Common String
Operations. R package version 1.5.1.
https://CRAN.R-project.org/package=stringr.
}
