% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePredictionStats.R
\name{create_seq_gap_heatmap}
\alias{create_seq_gap_heatmap}
\title{Create Heatmap}
\usage{
create_seq_gap_heatmap(seq_gaps_melted)
}
\arguments{
\item{seq_gaps_melted}{Melted dataframe of sequence coverage}
}
\description{
filler description
}
