% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPredictions.R
\name{create_sequence}
\alias{create_sequence}
\title{Create Sequence from DP table}
\usage{
create_sequence(dp, hsp_table, query_len)
}
\arguments{
\item{dp}{The dynamic programming table}

\item{hsp_table}{The table of hsps}

\item{query_len}{Length of the query}
}
\description{
Dummy Description
}
