% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignPredictions.R
\name{display_phylogeny}
\alias{display_phylogeny}
\title{Display a phylogeny of a gene set}
\usage{
display_phylogeny(predictions, gene_name)
}
\arguments{
\item{predictions}{description}

\item{gene_name}{description}
}
\description{
Sample description
}
