% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePredictionStats.R
\name{get_gene_coverage}
\alias{get_gene_coverage}
\title{Augment Dataset for Heatmap Generation (helper for gene_coverage_heatmap)}
\usage{
get_gene_coverage(predictions)
}
\arguments{
\item{predictions}{An S3 object of type GenePredictions, created by a
previous iteration of the build_predictions function in the package.}
}
\value{
Return a data frame with three columns:
\itemize{
 \item species name
 \item gene name
 \item coding sequence coverage of the predicted coding sequence for the
 species and gene
}
}
\description{
Given an input GenePredictions object containing predicted coding sequences,
calculate gene coverage (ie. prediction completeness) for each predicted
sequence. Store the calculations in a melted data frame for heatmap
generation.
}
\references{
Wickham H, Henry L. (2023). purrr: Functional Programming Tools. R package
version 1.0.2. https://CRAN.R-project.org/package=purrr.
}
