% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePredictionStats.R
\name{get_seq_gaps}
\alias{get_seq_gaps}
\title{HeatMap Generator Helper}
\usage{
get_seq_gaps(prediction_df)
}
\arguments{
\item{prediction_df}{Dataframe created by Build Predictions}
}
\description{
Filler Description
}
