% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePredictionStats.R
\name{get_unique_genes}
\alias{get_unique_genes}
\title{Get Unique Genes}
\usage{
get_unique_genes(prediction_df)
}
\arguments{
\item{prediction_df}{Database containing predictions, created by BuildPredictions}
}
\description{
filler description
}
