% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPredictions.R
\name{handle_overlap}
\alias{handle_overlap}
\title{HSP Overlap Merger (Helper for recursive_create)}
\usage{
handle_overlap(hsps, seq, p, i)
}
\arguments{
\item{hsps}{Dataframe of HSPs, sorted in ascending order by start position of
query segment}

\item{seq}{Currently built sequence by the recursive call in recursive_create}

\item{p}{Index of parent HSP}

\item{i}{Index of current HSP in consideration by recursive_create}
}
\value{
Returns seq and the subject segment of HSP i, merged together as a
string
}
\description{
Minimal algorithm for merging together two overlapping HSPs. If insertions
or deletions are present in one of the two overlapping segments, use the
subject sequence for the other one for prediction instead. I note that
despite assuming that relevant HSPs correspond loosely to exons, they may
extend extraneously into introns if the intron ends are similar to the
neighboring exon ends.
}
\references{
Wheeler, D. and Bhagwat, M. Blast QuickStart: Example-Driven Web-Based BLAST
Tutorial. In: Bergman NH, editor. Comparative Genomics: Volumes 1 and 2.
Totowa (NJ): Humana Press; 2007. Chapter 9.
https://www.ncbi.nlm.nih.gov/books/NBK1734/
}
