% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{my_function}
\alias{my_function}
\title{Oxygen tag test (title)}
\usage{
my_function(name1, name2, name3)
}
\arguments{
\item{name1}{first item to be printed if we keep typing and typing and}

\item{name2}{second item to be printed}

\item{name3}{test filepath}
}
\value{
nothing!you can also see what specifically they should look like
}
\description{
A test function that just prints out the name (description)
}
