% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputPredictions.R
\name{output_predictions_df}
\alias{output_predictions_df}
\title{Output predictions in data frame format}
\usage{
output_predictions_df(predictions)
}
\arguments{
\item{predictions}{An S3 object of type GenePredictions, produced by the
package's build_predictions function.}
}
\value{
Returns a data frame containing the predicted coding sequences. The
data frame has three string-containing columns: 'Gene', 'Species' and
'Sequence', corresponding to the gene name, species and sequence for a given
prediction. If no coding sequence prediction is available for a gene of a
particular species, no row for the gene-species combination is present.
}
\description{
Given an input nested list of gene predictions in GenePrediction format,
output the predictions as a data frame with three columns corresponding to
gene name, species and sequence.
}
\examples{

# Get gene predictions from sample RawBlastList dataset provided in the
# package
gene_predictions <- build_predictions(sample_raw_blast_list,
                                      max_overlap = 30,
                                      max_intron_length = 50000)

# Convert the output to a data frame containing the predicted coding
# sequences
output_predictions_data <- output_predictions_df(predictions)

}
\references{
Wickham H, Henry L. (2023). purrr: Functional Programming Tools. R package
version 1.0.2. https://CRAN.R-project.org/package=purrr.
}
