% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputPredictions.R
\name{output_predictions_fasta}
\alias{output_predictions_fasta}
\title{Output BLAST predictions in FASTA format}
\usage{
output_predictions_fasta(predictions, output_file = NULL)
}
\arguments{
\item{predictions}{An S3 object of type GenePredictions, produced by the
package's build_predictions function.}

\item{output_file}{The file to output the fasta-format predictions to. NOTE
that anything in the output_file, if it exists will be OVERWRITTEN by this
function.}
}
\value{
Returns NULL. Prints out the FASTA string representation of the
predicted sequence if no output file is provided; otherwise, redirects output
to the specified filepath.
}
\description{
Given an input nested list of gene predictions in GenePrediction format,
output the predictions in fasta format. The output can be directly sent
to a file, or simply printed to the R console.
}
\examples{

# Get gene predictions from sample RawBlastList dataset provided in the
# package
gene_predictions <- build_predictions(sample_raw_blast_list,
                                      max_overlap = 30,
                                      max_intron_length = 50000)

# Example 1: Output gene predictions in FASTA format to console as a string
\dontrun{

output_predictions_fasta(predictions = gene_predictions,
                         output_file = NULL)
}

# Example 2: Output gene predictions in FASTA format to a file
\dontrun{

file_path <- /path/to/file # make sure there is no existing info in it
output_predictions_fasta(predictions = gene_predictions,
                         output_file = file_path)

}
}
\references{
Wickham H, Henry L. (2023). purrr: Functional Programming Tools. R package
version 1.0.2. https://CRAN.R-project.org/package=purrr.
}
