% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseInput.R
\name{parse_json}
\alias{parse_json}
\title{Parse Input Title}
\usage{
parse_json(filename)
}
\arguments{
\item{filename}{file name of the BLAST XML/json input file to parse}
}
\description{
Parse Input Description
}
