% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseInput.R
\name{parse_multiple_BLAST_json}
\alias{parse_multiple_BLAST_json}
\title{Parse Multiple BLAST json}
\usage{
parse_multiple_BLAST_json(dir_path, raw_list = NULL)
}
\arguments{
\item{dir_path}{path to directory containing json files from different
species}

\item{raw_list}{previous raw carrier, if available}
}
\description{
description
}
