% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseInput.R
\name{validate_BLAST_JSON_file}
\alias{validate_BLAST_JSON_file}
\title{Validate BLAST JSON output file (Helper for parse_BLAST_json)}
\usage{
validate_BLAST_JSON_file(filename)
}
\arguments{
\item{filename}{File path to the BLAST JSON output file to validate.}
}
\value{
Return the contents of the BLAST JSON output file in R list format
as directly translated by the fromJSON() function from the rjson package.
The key-value 'dictionary' format of JSON is preserved in the list.
}
\description{
Checks to see if the provided BLAST output file is 1) in .json format, and
2) follows the conventions of results from NCBI's BLAST search tool. See the
example BLAST output files in the inst/extdata directory for more
information. In general, each BLAST output file must have the following
attribute structure:
\itemize{
 \item An overarching attribute 'BlastOutput2'. The value of 'BlastOutput2'
 is a list of 'reports', each containing results for a query.
 \item Within a 'report', the query title and length can be accessed by
 traversing the attributes: 'report' -> 'results' -> 'search'.
 Their respective attribute names are 'query_title' and 'query_len'
 \item The 'hits' attribute can also be accessed this way, and contains
 a list of hits each representing a collection of HSPs for a single subject
 sequence.
}
}
\references{
Couture-Beil A (2024). rjson: JSON for R. R package
version 0.2.23.
https://CRAN.R-project.org/package=rjson.

Li J. (2020). _configr: An Implementation of Parsing and
Writing Configuration File (JSON/INI/YAML/TOML)_. R
package version 0.3.5,
https://CRAN.R-project.org/package=configr.

Wheeler, D. and Bhagwat, M. Blast QuickStart: Example-Driven Web-Based BLAST
Tutorial. In: Bergman NH, editor. Comparative Genomics: Volumes 1 and 2.
Totowa (NJ): Humana Press; 2007. Chapter 9.
https://www.ncbi.nlm.nih.gov/books/NBK1734/

Wickham H, François R, Henry L, Müller K, Vaughan D. (2023). _dplyr: A
Grammar of Data Manipulation_. R package version 1.1.4.
https://CRAN.R-project.org/package=dplyr.

Wickham H, Hester J, Bryan J. (2024). _readr: Read
Rectangular Text Data_. R package version 2.1.5.
https://CRAN.R-project.org/package=readr.
}
