% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisplayPhylogeny.R
\name{display_phylogeny}
\alias{display_phylogeny}
\title{Display Phylogeny for a Single Gene}
\usage{
display_phylogeny(predictions, gene_name)
}
\arguments{
\item{predictions}{S3 object of class GenePredictions produced by an
iteration of the 'build_predictions' function in this package.}

\item{gene_name}{Name of gene to construct phylogeny from. As a requirement,
there must be at least coding sequence for the specified gene in
'predictions'.}
}
\value{
A plot of type "phylogram", generated by the base graphics package.
The user should use this plot as a quick way to discern similarity between
the predicted sequences of different species, rather than a real phylogenetic
tree.
}
\description{
Given an input nested list of predicted coding sequences in GenePredictions
format and an user-specified gene, construct a phylogeny from
predicted coding sequences for that single gene. After isolating the
relevant coding sequences for the gene from the
GenePredictions object, the sequences are then aligned with ClustalW,
a multiple-sequence alignment algorithm. Finally, a phylogeny is
inferred via the Unweighted Pair Group Method with Arithmetic Mean (UPGMA)
algorithm. UPGMA assumes that:
\itemize{
\item all taxa evolve at a constant rate, and that
\item all species also have the same mutation rate
}
Note that these assumptions are breached quite often in reality. Still,
although other methods like maximum-likelihood-based approaches have greater
accuracy, they take far longer to execute. The user should therefore treat
this function as a quick way to visually discern similarity between the
predicted coding sequences, rather than a real reflection of phylogenetic
relationships. If species appear closely related in the phylogenies produced,
the user should merely assume that their predicted sequences are closest to
each other in comparison to the sequences predicted for other species.
}
\examples{

# make gene predictions from sample RawBlastList data provided in package
predictions <- build_predictions(sample_raw_blast_list,
                                 max_overlap = 30,
                                 max_intron_length = 50000)

# Example 1: Display phylogeny based on predicted sequences for RHO
\dontrun{

display_phylogeny(predictions = predictions, gene_name = "RHO")

}

# Example 2: Display phylogeny based on predicted sequences for IL6
\dontrun{

display_phylogeny(predictions = predictions, gene_name = "IL6")
# Note that the species that the algorithm failed to predict sequences for
# (P. marinus and D. rerio) are not in the phylogeny.

}

}
\references{
Bodenhofer U, Bonatesta E, Horejs-Kainrath C, and Hochreiter S. (2015).
msa: an R package for multiple sequence alignment. Bioinformatics
31(24):3997-9999. DOI: 10.1093/bioinformatics/btv176.

Pagès H, Aboyoun P, Gentleman R, DebRoy S (2024). Biostrings: Efficient
manipulation of biological strings. doi:10.18129/B9.bioc.Biostrings.

Schlie K, Potts AJ, Morrison DA, Grimm, GW. (2017), Intertwining
phylogenetic trees and networks. Methods in Ecology and Evolution, 8:
1212--1220. doi:10.1111/2041-210X.12760

Sokal R, Michener CD. (1958). A Statistical Method for Evaluating
Systematic Relationships. University of Kansas Scientific Bulletin
28:1409-1438.

Wickham H, François R, Henry L, Müller K, Vaughan D. (2023). dplyr: A
Grammar of Data Manipulation. R package version 1.1.4.
https://CRAN.R-project.org/package=dplyr.

Wickham H, Henry L. (2023). purrr: Functional Programming Tools. R package
version 1.0.2. https://CRAN.R-project.org/package=purrr.
}
