% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePredictionStats.R
\name{gene_coverage_heatmap}
\alias{gene_coverage_heatmap}
\title{Create Heatmap}
\usage{
gene_coverage_heatmap(predictions)
}
\arguments{
\item{melted_coverages}{Melted dataframe of sequence coverage}
}
\description{
filler description
}
