% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePredictionStats.R
\name{get_gene_coverage}
\alias{get_gene_coverage}
\title{Augment Dataset for Heatmap Generation (helper for gene_coverage_heatmap)}
\usage{
get_gene_coverage(predictions)
}
\arguments{
\item{predictions}{Data frame created by Build Predictions}
}
\description{
Given an input GenePredictions object containing predicted coding sequences
}
