% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPredictions.R
\name{handle_overlap}
\alias{handle_overlap}
\title{Merge Overlap Helper}
\usage{
handle_overlap(hsps, seq, p, i)
}
\arguments{
\item{hsps}{Table of HSPs}

\item{seq}{Currently built sequence}

\item{p}{Index of parent sequence}

\item{i}{Index of current sequence}
}
\description{
Placeholder description
}
