% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPredictions.R
\name{is_compatibile}
\alias{is_compatibile}
\title{Helper function for building predictions.}
\usage{
is_compatibile(hsp_table, j, i)
}
\arguments{
\item{hsp_table}{a sorted table of HSPs}

\item{j}{index of HSP with lower query start}

\item{i}{index of HSP with high query start}
}
\description{
Determines if two HSPs are compatible
}
