% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseInput.R
\name{parse_BLAST_json}
\alias{parse_BLAST_json}
\title{Parse Input Title}
\usage{
parse_BLAST_json(filename, species)
}
\arguments{
\item{filename}{File name of the BLAST XML/json input file to parse}

\item{species}{Species from which the subject genome is from.}
}
\description{
Parse Input Description
}
