% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseInput.R
\name{parse_multiple_BLAST_json}
\alias{parse_multiple_BLAST_json}
\title{Parse Multiple BLAST JSON Output Files}
\usage{
parse_multiple_BLAST_json(dir_path, raw_blast_list = NULL)
}
\arguments{
\item{dir_path}{Path to directory containing BLAST output .json files. Each
file must be the BLAST output from coding sequences as queries and
subject genomic sequences from a single species DIFFERENT from other runs.
The name of each .json file must be the species name for the subject set of
genomic sequences blasted against.}

\item{raw_blast_list}{Previous RawBlastList generated by the package. Use to
concatenate the newly generated RawBlastList and new one together. Default is
NULL.}
}
\value{
Returns an S3 object of type RawBlastList. See parse_BLAST_json for a
more detailed description.
}
\description{
Given a directory containing multiple BLAST output files in .json format,
(where each BLAST run was conducted on coding sequences as queries and set
of subject genomic sequences for a species different from other runs),
generate a RawBlastList storing the data (see parse_BLAST_json for a more
detailed description). Assumptions are the same as parse_BLAST_json.
}
\examples{

# "sample_BLAST_output_dir" contains BLAST output for two different
# instances where a rhodopsin query is used against a subject set of Orca
# genomic sequences (results titled O_orca.json) and when it is used against
# a subject set of Dolphin genomic sequences (results titled
# T_truncatus.json)

BLAST_out_directory <- system.file("extdata", "example_raw_output_directory",
                                   package="euPredictR")

# EXAMPLE: Convert data in BLAST_out_directory to RawBlastList format

raw_blast_list1 <- parse_multiple_BLAST_json(dir_path = BLAST_out_directory,
                                             raw_blast_list = NULL)

}
\references{
R Core Team. (2024). R: A Language and Environment for Statistical Computing.
R Foundation for Statistical Computing,
Vienna, Austria. https://www.R-project.org/.
}
