% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPredictions.R
\name{recursive_create}
\alias{recursive_create}
\title{Recursive Helper}
\usage{
recursive_create(dp, hsps, i)
}
\arguments{
\item{dp}{The dynamic programming table}

\item{i}{index of the last HSP in the sequence to create}

\item{hsp_table}{The table of hsps}
}
\description{
Dummy Description
}
