% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseInput.R
\name{remove_gaps_if_equal}
\alias{remove_gaps_if_equal}
\title{Check if number of gaps is equal and remove if so}
\usage{
remove_gaps_if_equal(hsp_data)
}
\arguments{
\item{hsp_data}{Table containing HSP data}
}
\description{
Dummy Description
}
