% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseInput.R
\name{remove_gaps_if_equal}
\alias{remove_gaps_if_equal}
\title{Remove Gaps in High Scoring Pair (HSP) Segment (Helper for parse_BLAST_json)}
\usage{
remove_gaps_if_equal(hsp_data)
}
\arguments{
\item{hsp_data}{Data frame containing high-scoring pair (HSP) data. Each row
is an HSP, and the columns are HSP attributes.}
}
\value{
Returns the altered HSP data frame, now with gaps removed if both
both the query and subject have an equal number of them.
}
\description{
Sometimes, high-scoring pairs (composed of a query segment and a subject
segment) contain gaps from alignment. When conducting predictions, it is
often useful to preserve these gaps to easily track 'missing' or 'extra'
nucleotides in the subject segment. However, if both the query and subject
have an equal number of gaps, removing them only shifts the alignment (there
aren't any "missing" or "extra" nucleotides), and helps remove any
frameshifting that might otherwise occur in the final gene prediction.
}
\references{
Wheeler, D. and Bhagwat, M. Blast QuickStart: Example-Driven Web-Based BLAST
Tutorial. In: Bergman NH, editor. Comparative Genomics: Volumes 1 and 2.
Totowa (NJ): Humana Press; 2007. Chapter 9.
https://www.ncbi.nlm.nih.gov/books/NBK1734/

Wickham H. (2023). stringr: Simple, Consistent Wrappers for Common String
Operations. R package version 1.5.1,
<https://CRAN.R-project.org/package=stringr>.
}
