% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEuPredictR.R
\name{run_euPredictR}
\alias{run_euPredictR}
\title{Launch Shiny App for euPredictR}
\usage{
run_euPredictR()
}
\value{
No return value but open up a Shiny page
}
\description{
A function that launches the Shiny app for euPredictR.
}
\examples{
\dontrun{
# Runs the shiny application for euPredictR
euPredictR::run_euPredictR()

}


}
